Move localization to ui folder
